populate providers overview card
