+++ packages/api/src/router/resource-provider-page/router.ts
@@ -1,0 +1,6 @@
+import { createTRPCRouter } from "../../trpc";
+import { overviewRouter } from "./overview/router";
+
+export const resourceProviderPageRouter = createTRPCRouter({
+  overview: overviewRouter,
+});
